fixed view.js into components
