+++ assets/js/components/index.js
@@ -1,0 +1,7 @@
+// components/index.js - Convenient export of all components
+
+export * from './NavigationComponent.js';
+export * from './UIStateManager.js';
+export * from './ProductComponent.js';
+export * from './CategoryComponent.js';
+export * from './DataLoader.js';
